+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/ai_assistant/knowledge_base/knowledge_base_add_semantic_text_field_migration.spec.ts
@@ -18,7 +18,7 @@ import {
   deleteInferenceEndpoint,
   setupKnowledgeBase,
   waitForKnowledgeBaseReady,
-} from './helpers';
+} from '../utils/knowledge_base';
 
 interface InferenceChunk {
   text: string;
